backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pdfplumber==0.11.4
python-docx==1.1.0
openai==1.40.3
httpx==0.27.2
python-dotenv==1.0.1
SQLAlchemy==2.0.34
pydantic-settings==2.5.2
